Added charging efficiency
This model assumes constant charging and discharging rates for each hour.
If the power input from solar is greater than the power output to the
load, then the system is charging. This efficiency is used. The
efficiency only applies to the power difference as this is what goes to
the battery.
If the power input from solar is less than the power output to the
load, then the system is discharging. This value does not apply in
discharging.